+++ R/load_fotmob.R
@@ -1,0 +1,67 @@
+#' Load fotmob match details
+#'
+#' Loading version of \code{fotmob_get_match_details}, but for all seasons
+#' List columns have been unnested in a sensible manner. Note that fotmob
+#' only has match details going back to the 2020-21 season for most leagues.
+#'
+#' @inheritParams fotmob_get_league_matches
+#'
+#' @return returns a dataframe of league matches
+#'
+#' @importFrom purrr possibly map_dfr
+#' @importFrom rlang maybe_missing
+#' @importFrom cli cli_alert
+#'
+#' @examples
+#' \dontrun{
+#' try({
+#' # one league
+#' load_fotmob_match_details(
+#'   country = "ENG",
+#'   league_name = "Premier League"
+#' )
+#'
+#' ## this is the same output format as the following
+#' fotmob_get_match_details(match_id = 3411352)
+#'
+#' # one league, by id
+#' load_fotmob_match_details(league_id = 47)
+#'
+#' # multiple leagues (could also use ids)
+#' load_fotmob_match_details(
+#'   country =     c("ENG",            "ESP"   ),
+#'   league_name = c("Premier League", "LaLiga")
+#' )
+#' })
+#' }
+#' @export
+load_fotmob_match_details <- function(country, league_name, league_id, cached = TRUE) {
+
+  fotmob_urls <- .fotmob_get_league_ids(
+    cached = cached,
+    country = rlang::maybe_missing(country, NULL),
+    league_name = rlang::maybe_missing(league_name, NULL),
+    league_id = rlang::maybe_missing(league_id, NULL)
+  )
+
+  urls <- sprintf(
+    "https://github.com/JaseZiv/worldfootballR_data/blob/master/data/fotmob_match_details/%s_match_details.rds?raw=true",
+    fotmob_urls$id
+  )
+
+  fp <- purrr::possibly(
+    .file_reader,
+    quiet = FALSE,
+    otherwise = data.frame()
+  )
+
+  res <- purrr::map_dfr(urls, fp)
+
+  if(nrow(res) == 0) {
+    cli::cli_alert("Data not loaded. Please check parameters")
+  } else {
+    ## when there are multiple data sets loaded in, seems like this is the attribute for the first
+    cli::cli_alert("Data last updated {attr(res, 'scrape_timestamp')} UTC")
+  }
+  res
+}